+++ build-support/custom-modules/ns3-compiler-workarounds.cmake
@@ -23,21 +23,12 @@ include(CheckCXXSourceCompiles)
 # link it manually. https://en.cppreference.com/w/cpp/filesystem
 check_cxx_source_compiles(
   "
-  #ifdef __has_include
-    #if __has_include(<filesystem>)
-      #include <filesystem>
-      namespace fs = std::filesystem;
-    #elif __has_include(<experimental/filesystem>)
-      #include <experimental/filesystem>
-      namespace fs = std::experimental::filesystem;
-    #else
-      #error \"No support for filesystem library\"
-    #endif
-  #endif
+  #include <filesystem>
+
   int main()
   {
     std::string path = \"/\";
-    return !fs::exists(path);
+    return !std::filesystem::exists(path);
   }
   "
   FILESYSTEM_LIBRARY_IS_LINKED